add option to hide orbits, modify electron speed. Fix firefox mouseleave bug
added the option to hide the orbits in a similar way to what was done with the nucleus in internal mode. Trajectory constant for the speed computation is also now modifiable. Moreover a reset button has been added, it resets the camera and model visualization parameters. The fix involving firefox was fixed by ignoring onmouseleave events not triggered by the panels